+++ Beats/SourceCode/Render/RenderCache/LineListCache.h
@@ -7,7 +7,7 @@ class CLineListCache : public CRenderCacheBase
 {
     typedef CRenderCacheBase super;
 public:
-    CLineListCache(size_t maxVertexCount, IDirect3DDevice9* m_pRenderDevice, long vertexFVF, long vertexSize);
+    CLineListCache(uint32_t maxVertexCount, IDirect3DDevice9* m_pRenderDevice, long vertexFVF, long vertexSize);
     virtual ~CLineListCache();
 
     virtual void AddIndex();
