+++ Beats/SourceCode/Render/RenderCache/RenderCacheBase.h
@@ -9,7 +9,7 @@ class Function;
 class CRenderCacheBase
 {
 public:
-    CRenderCacheBase(size_t maxVertexCount, IDirect3DDevice9* m_pRenderDevice, long vertexFVF, long vertexSize,  ERenderCacheType type);
+    CRenderCacheBase(uint32_t maxVertexCount, IDirect3DDevice9* m_pRenderDevice, long vertexFVF, long vertexSize,  ERenderCacheType type);
     virtual ~CRenderCacheBase();
     
     virtual void AddIndex() = 0;
@@ -27,9 +27,9 @@ public:
 
 protected:
     DWORD* m_pCurIndex;
-    size_t m_vertexCounter;
-    size_t m_maxVertexCount;
-    size_t m_functionCounter;
+    uint32_t m_vertexCounter;
+    uint32_t m_maxVertexCount;
+    uint32_t m_functionCounter;
     D3DPRIMITIVETYPE m_prmitiveType;
     ERenderCacheType m_type;
     long m_vertexFVF;
@@ -38,7 +38,7 @@ protected:
     IDirect3DDevice9* m_pRenderDevice;
     IDirect3DVertexBuffer9* m_pVertexBuffer;
     IDirect3DIndexBuffer9* m_pIndexBuffer;
-    std::vector<std::pair<size_t, Function<void(void)>>> m_functionCache;
+    std::vector<std::pair<uint32_t, Function<void(void)>>> m_functionCache;
 };
 
 #endif
