+++ Beats/SourceCode/Render/RenderManager.h
@@ -36,7 +36,7 @@ private:
     bool m_bInit;
     std::vector<CRenderCacheBase*> m_caches;
     //Never erase element in m_unitsContainer, we recycle use them with mark and sweep
-    std::vector<size_t> m_recycleUnitId;
+    std::vector<uint32_t> m_recycleUnitId;
     std::vector<CRenderUnit*> m_unitsContainer;
     D3DPRESENT_PARAMETERS m_d3dPresentParam;
 };
