+++ Beats/SourceCode/Sound/SoundObject/DSMP3Object.cpp
@@ -90,7 +90,7 @@ bool CDSMP3Object::LoadDSData(unsigned int start, unsigned int count)
 
         LPVOID aptr1 = NULL, aptr2 = NULL;
         DWORD abyte1 = NULL, abyte2 = NULL;
-        size_t outsize = 0;
+        uint32_t outsize = 0;
         // we should promise no overflow in Lock method, so we don't care about aptr2
         if(m_pDSBuffer->Lock(start, count, &aptr1, &abyte1, &aptr2, &abyte2, 0) == DS_OK)
         {
@@ -104,7 +104,7 @@ bool CDSMP3Object::LoadDSData(unsigned int start, unsigned int count)
             {
                 //clear useless data to avoid noise when play to the end.
                 long resetDataBytes = bufferPieceSize >= (unsigned long)outsize ? bufferPieceSize - (unsigned long)outsize : m_bufferSize - (unsigned long)outsize;
-                memset((LPVOID)((size_t)aptr1 + outsize), 0, resetDataBytes);
+                memset((LPVOID)((uint32_t)aptr1 + outsize), 0, resetDataBytes);
 
                 m_EOFPosInBuffer = start + outsize - 1;
                 DWORD playPos = 0;
